Fix compiler warnings: add explicit casts for size_t to int conversions
- cString.h: cast strlen() results to int in operator= and operator+=
- cOrganism.h: cast donating_lineages.size() to int
- cGermline.h: cast m_germline.size() to unsigned int
- cTaskLib.h: cast m_strings.size() to int
- PopulationActions.cc: fix type conversions (int redundancy, double val, cast count_if result)

+++ avida-core/source/main/cGermline.h
@@ -41,7 +41,7 @@ public:
     
 	Genome& GetLatest() { assert(m_germline.size() > 0); return m_germline.back(); }
 	void Add(const Genome& genome) { m_germline.push_back(genome); }
-	unsigned int Size() const { return m_germline.size(); }
+	unsigned int Size() const { return static_cast<unsigned int>(m_germline.size()); }
 };
 
 #endif
